feat(react)!: flatten sdk-react imports
