+++ apps/kitchensink-react/src/DocumentCollection/DocumentPreview.tsx
@@ -1,5 +1,5 @@
 import {DocumentHandle} from '@sanity/sdk'
-import {usePreview} from '@sanity/sdk-react/hooks'
+import {usePreview} from '@sanity/sdk-react'
 import {Suspense, useRef} from 'react'
 import {ErrorBoundary} from 'react-error-boundary'
