+++ apps/kitchensink-react/src/DocumentCollection/DocumentProjectionRoute.tsx
@@ -1,5 +1,5 @@
 import {DocumentHandle} from '@sanity/sdk'
-import {usePaginatedList, useProjection} from '@sanity/sdk-react/hooks'
+import {usePaginatedList, useProjection} from '@sanity/sdk-react'
 import {Box, Button, Card, Flex, Label, Spinner, Stack, Text, TextInput} from '@sanity/ui'
 import {JSX, ReactNode, Suspense, useRef, useState} from 'react'
 import {ErrorBoundary} from 'react-error-boundary'
