--- packages/react/src/_exports/components.ts
@@ -1,3 +1,0 @@
-export {AuthBoundary} from '../components/auth/AuthBoundary'
-export {SanityApp, type SanityAppProps} from '../components/SanityApp'
-export {SDKProvider, type SDKProviderProps} from '../components/SDKProvider'
