--- packages/react/src/_exports/context.ts
@@ -1,2 +1,0 @@
-export type {SanityProviderProps} from '../context/SanityProvider'
-export {SanityProvider} from '../context/SanityProvider'
