+++ packages/react/src/_exports/index.ts
@@ -1,1 +1,58 @@
+export {AuthBoundary} from '../components/auth/AuthBoundary'
+export {SanityApp, type SanityAppProps} from '../components/SanityApp'
+export {SDKProvider, type SDKProviderProps} from '../components/SDKProvider'
+export type {SanityProviderProps} from '../context/SanityProvider'
+export {SanityProvider} from '../context/SanityProvider'
+export {useAuthState} from '../hooks/auth/useAuthState'
+export {useAuthToken} from '../hooks/auth/useAuthToken'
+export {useCurrentUser} from '../hooks/auth/useCurrentUser'
+export {useHandleCallback} from '../hooks/auth/useHandleCallback'
+export {useLoginUrls} from '../hooks/auth/useLoginUrls'
+export {useLogOut} from '../hooks/auth/useLogOut'
+export {useClient} from '../hooks/client/useClient'
+export {
+  type FrameConnection,
+  type FrameMessageHandler as MessageHandler,
+  useFrameConnection,
+  type UseFrameConnectionOptions,
+} from '../hooks/comlink/useFrameConnection'
+export {useManageFavorite} from '../hooks/comlink/useManageFavorite'
+export {useRecordDocumentHistoryEvent} from '../hooks/comlink/useRecordDocumentHistoryEvent'
+export {
+  useWindowConnection,
+  type UseWindowConnectionOptions,
+  type WindowConnection,
+  type WindowMessageHandler,
+} from '../hooks/comlink/useWindowConnection'
+export {useSanityInstance} from '../hooks/context/useSanityInstance'
+export {useDatasets} from '../hooks/datasets/useDatasets'
+export {useApplyActions} from '../hooks/document/useApplyActions'
+export {useDocument} from '../hooks/document/useDocument'
+export {useDocumentEvent} from '../hooks/document/useDocumentEvent'
+export {useDocumentSyncStatus} from '../hooks/document/useDocumentSyncStatus'
+export {useEditDocument} from '../hooks/document/useEditDocument'
+export {usePermissions} from '../hooks/document/usePermissions'
+export {
+  type InfiniteList,
+  type InfiniteListOptions,
+  useInfiniteList,
+} from '../hooks/infiniteList/useInfiniteList'
+export {
+  type PaginatedList,
+  type PaginatedListOptions,
+  usePaginatedList,
+} from '../hooks/paginatedList/usePaginatedList'
+export {
+  usePreview,
+  type UsePreviewOptions,
+  type UsePreviewResults,
+} from '../hooks/preview/usePreview'
+export {useProjection} from '../hooks/projection/useProjection'
+export {useProject} from '../hooks/projects/useProject'
+export {type ProjectWithoutMembers, useProjects} from '../hooks/projects/useProjects'
+export {useQuery} from '../hooks/query/useQuery'
+export {useUsers, type UseUsersParams, type UseUsersResult} from '../hooks/users/useUsers'
 export {REACT_SDK_VERSION} from '../version'
+export {type DatasetsResponse, type SanityProject, type SanityProjectMember} from '@sanity/client'
+export {type CurrentUser, type DocumentHandle} from '@sanity/sdk'
+export {type SanityDocument, type SortOrderingItem} from '@sanity/types'
